apps/users/serializers.py
from rest_framework import serializers
from django.contrib.auth import get_user_model
from apps.users.models import User

class UserSerializer(serializers.ModelSerializer):
    class Meta:
        model = User
        fields = ['id', 'username', 'email', 'phone', 'first_name', 'last_name']
