requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-jazzmin==3.0.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
pytz==2024.2
PyYAML==6.0.2
setuptools==75.1.0
sqlparse==0.5.1
uritemplate==4.1.1
